Add the property 'onAllChecked' to be set when passing the config object of the table
A boolean will me passed as a parameters informing that: all the items on the current page are "checked" (TRUE); or at least one item is not (FALSE). The function is called whenever that status change. So this function is only called when all the items get checked (not necessarily at once), or, if they are all checked, when one is un-checked.