Change default emptyTableTemplate's language to english, adjust clearData function and add clearTable function.
The function clearData now is reserved for clearing the table's data making it empty, and as a default behavior when the table is empty it will display the message "No item found" (the message's html can be overriden on the configs). And clearTable will be the function to make the table list null, and that won't show any message, just the table headers.